Add view all function and few improvements in Trains class

+++ RailwayReservationSystem/Train.cs
@@ -19,30 +19,48 @@ namespace RailwayReservationSystem
         //Navigation Entries
         public ICollection<Schedule> Schedule { get; set; }
 
-        public void AddNew(int capacity)
+        public Train()
+        {
+
+        }
+        public Train(int capacity)
         {
             using var context = new ApplicationDbContext();
 
-            var NewTrain = new Train
-            {
-                Capacity = capacity,
-                SeatsAvailable = capacity,
-                SeatsBooked = 0
-            };
+            Capacity = capacity;
+            SeatsAvailable = capacity;
+            SeatsBooked = 0;
 
-            context.Trains.Add(NewTrain);
+            context.Trains.Add(this);
 
             context.SaveChanges();
         }
+
         public void View()
         {
             Console.WriteLine("Train No: T-" + this.TrainNo);
             Console.WriteLine("Seating Capacity: " + this.Capacity);
-            Console.WriteLine("Seats Booked: " + this.SeatsBooked);
             Console.WriteLine("Seats Available: " + this.SeatsAvailable);
+            Console.WriteLine("Seats Booked: " + this.SeatsBooked);
             Console.WriteLine(new String('-', 30));
         }
-        public void Update(int trainNo)
+        public void ViewAll()
+        {
+            using var context = new ApplicationDbContext();
+
+            var trains = from t in context.Trains
+                         select t;
+
+            foreach(var train in trains)
+            {
+                Console.WriteLine("Train No: T-" + train.TrainNo);
+                Console.WriteLine("Seating Capacity: " + train.Capacity);
+                Console.WriteLine("Seats Available: " + train.SeatsAvailable);
+                Console.WriteLine("Seats Booked: " + train.SeatsBooked);
+                Console.WriteLine(new String('-', 30));
+            }
+        }
+        public void Update(int trainNo, int newcapacity)
         {
             using var context = new ApplicationDbContext();
 
@@ -50,10 +68,16 @@ namespace RailwayReservationSystem
                          where t.TrainNo == trainNo
                          select t).FirstOrDefault();
 
-            Console.WriteLine("Enter Updated Train Capacity: ");
-
-            train.Capacity = int.Parse(Console.ReadLine());
-            train.SeatsAvailable = train.Capacity - train.SeatsBooked;
+            if (train != null)
+            {
+                train.Capacity = newcapacity;
+                train.SeatsAvailable = train.Capacity - train.SeatsBooked;
+            }
+            else
+            {
+                Console.WriteLine("Train does not exist");
+                return;
+            }
 
             context.SaveChanges();
         }
@@ -65,7 +89,15 @@ namespace RailwayReservationSystem
                          where t.TrainNo == trainNo
                          select t).FirstOrDefault();
 
-            context.Trains.Remove(train);
+            if (train != null)
+            {
+                context.Trains.Remove(train);
+            }
+            else
+            {
+                Console.WriteLine("Train does not exist");
+                return;
+            }
 
             context.SaveChanges();
         }
